Add files via upload
Our goal is to determine what factors are most useful in determining whether an animal will be adopted. This will allow shelters to predict which animals will be adopted.  Also, if we are able to determine which locations are mostly to have adoptions of each type of animal, then shelters can transfer pets to where they are most likely to be adopted.  We also want to predict age based on various attributes. This will allow shelters to better estimate the age of pets that are brought in.